feat:기본/추가정보 입력 페이지 분리
