+++ src/pages/Personal/my/info/ExtraInfo.tsx
@@ -1,0 +1,5 @@
+const ExtraInfo = () => {
+  return <div>추가정보</div>;
+};
+
+export default ExtraInfo;
