+++ src/pages/Personal/my/info/PersonalInfo.tsx
@@ -1,121 +1,41 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import Topbar from '../../../../shared/components/topbar/Topbar';
 import clsx from 'clsx';
-
-interface info {
-  name: string;
-  id: string;
-  password: string;
-  age: number;
-  phone: string;
-  address: {
-    main: string;
-    detail: string;
-  };
-  job: string;
-  period: string;
-}
-
-const dummyInfo: info = {
-  name: '김영희',
-  id: 'youngid',
-  password: '1234',
-  age: 63,
-  phone: '010-0000-0000',
-  address: {
-    main: '대지로 49',
-    detail: '203동 105호',
-  },
-  job: '주부',
-  period: '10년 이상',
-};
-
-const infoData = [
-  { label: '이름', value: dummyInfo.name },
-  { label: 'id', value: dummyInfo.id },
-  { label: '비밀번호', value: '*****' },
-  { label: '나이', value: dummyInfo.age },
-  { label: '연락처', value: dummyInfo.phone },
-  {
-    label: '거주지',
-    value: `${dummyInfo.address.main}\n${dummyInfo.address.detail}`,
-  },
-  { label: '직무', value: dummyInfo.job },
-  { label: '기간', value: dummyInfo.period },
-];
+import Topbar from '../../../../shared/components/topbar/Topbar';
+import BasicInfo from './BasicInfo';
+import ExtraInfo from './ExtraInfo';
 
 const PersonalInfo = () => {
-  const navigate = useNavigate();
   const [isClicked, setIsClicked] = useState<'basic' | 'extra'>('basic');
 
   return (
     <div>
       <Topbar>
-        <div className="flex flex-col justify-center items-center text-[16px] font-[500]">
-          <div className="text-center font-[pretendard JP] font-[600] text-[20px] text-[#747474] py-[10px]">
-            개인정보
-          </div>
-          <div className="flex justify-center items-center gap-[10px] font-[Pretendard JP] text-black">
-            <button
-              type="button"
-              className={clsx(
-                isClicked === 'basic' ? 'bg-[#ECF6F2]' : '',
-                'w-[144px] h-[45px] rounded-[8px] border-[1.3px] border-[#08D485]'
-              )}
-              onClick={() => setIsClicked('basic')}
-            >
-              기본 정보
-            </button>
-            <button
-              type="button"
-              className={clsx(
-                isClicked === 'basic' ? '' : 'bg-[#ECF6F2]',
-                'w-[144px] h-[45px] rounded-[8px] border-[1.3px] border-[#08D485]'
-              )}
-              onClick={() => setIsClicked('extra')}
-            >
-              추가 정보
-            </button>
-          </div>
-          <div className="py-[26px]">
-            {infoData.map(({ label, value }) => (
-              <div
-                key={label}
-                className="w-full flex justify-center gap-[30px] text-black leading-[2.8]"
-              >
-                <span className="w-[140px] pl-[30px] flex justify-end items-center text-[18px] text-[#414141]">
-                  {label}
-                </span>
-                <span className="w-full flex justify-center items-center text-[#C2C2C2] text-[16px]">
-                  {label === '비밀번호' ? (
-                    <button
-                      className="w-[170px] h-[40px] rounded-[8px] font-[Pretendard] text-white bg-[#08D485]"
-                      onClick={() => navigate('/personal/my/info/password')}
-                    >
-                      수정
-                    </button>
-                  ) : typeof value === 'string' && value.includes('\n') ? (
-                    <div className="text-center leading-tight">
-                      {value.split('\n').map((line, idx) => (
-                        <p key={idx}>{line}</p>
-                      ))}
-                    </div>
-                  ) : (
-                    value
-                  )}
-                </span>
-              </div>
-            ))}
-          </div>
+        <div className="text-center font-[pretendard JP] font-[600] text-[20px] text-[#747474] py-[10px]">
+          개인정보
+        </div>
+        <div className="flex justify-center items-center gap-[10px] font-[Pretendard JP] font-[500] text-black text-[16px]">
           <button
             type="button"
-            className="w-[309px] h-[45px] rounded-[8px] bg-[#08D485] text-white"
-            onClick={() => navigate('/personal/my/info/edit')}
+            className={clsx(
+              isClicked === 'basic' ? 'bg-[#ECF6F2]' : '',
+              'w-[144px] h-[45px] rounded-[8px] border-[1.3px] border-[#08D485]'
+            )}
+            onClick={() => setIsClicked('basic')}
           >
-            수정
+            기본 정보
+          </button>
+          <button
+            type="button"
+            className={clsx(
+              isClicked === 'basic' ? '' : 'bg-[#ECF6F2]',
+              'w-[144px] h-[45px] rounded-[8px] border-[1.3px] border-[#08D485]'
+            )}
+            onClick={() => setIsClicked('extra')}
+          >
+            추가 정보
           </button>
         </div>
+        {isClicked === 'basic' ? <BasicInfo /> : <ExtraInfo />}
       </Topbar>
     </div>
   );
