+++ src/pages/Personal/my/info/PersonalInfoEdit.tsx
@@ -3,7 +3,7 @@ import Topbar from '../../../../shared/components/topbar/Topbar';
 const PersonalInfoEdit = () => {
   return (
     <div>
-      <Topbar></Topbar>
+      <Topbar>개인정보수정페이지</Topbar>
     </div>
   );
 };
