+++ src/pages/Personal/my/info/PersonalInfoEditPassword.tsx
@@ -1,5 +1,11 @@
+import Topbar from '../../../../shared/components/topbar/Topbar';
+
 const PersonalInfoEditPassword = () => {
-  return <></>;
+  return (
+    <div>
+      <Topbar>비밀번호수정페이지</Topbar>
+    </div>
+  );
 };
 
 export default PersonalInfoEditPassword;
